requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncio==3.4.3
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
DateTime==5.5
Deprecated==1.2.14
durationpy==0.9
fastapi==0.115.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
kubernetes==31.0.0
logging==0.4.9.6
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
numpy==2.1.2
oauthlib==3.2.2
onnxruntime==1.20.0
opentelemetry-api>=1.28.0
opentelemetry-exporter-otlp-proto-common>=1.28.0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0
opentelemetry-instrumentation>=0.49b0
opentelemetry-instrumentation-asgi>=0.49b0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-proto>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-semantic-conventions>=0.49b0
opentelemetry-util-http>=0.49b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
posthog==3.7.0
proto-plus==1.25.0
protobuf>=5.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
tokenizers>=0.21.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
zipp==3.20.2
zope.interface==7.1.1
